Add timestamp to incoming PTP (over UDP) packets
TRex running without `--software` mode enabled but in service mode can process
PTP over UDP packets in RX.  But DPDK layer did not timestamp those packets.
This patch kind of implements timestamping using DPDK's rx-callbacks.  As this
differs from the original hardware timestamping some modifications were required
in RXTimesync class as well.

In order to have rx-callbacks available, RTE_ETHDEV_RXTX_CALLBACKS needs to be
enabled in DPDK's configuration.

+++ src/os_time.h
@@ -83,7 +83,7 @@ static inline hr_time_t    os_get_hr_freq(void){
 static inline hr_time_t get_rte_epoch_nanoseconds(uint8_t port) {
     struct timespec res{0, 0};
     if(rte_eth_timesync_read_time(port, &res) == 0) {
-        return res.tv_sec * 1000 * 1000 * 1000 + res.tv_nsec;
+        return ((uint64_t) res.tv_sec * 1000 * 1000 * 1000) + res.tv_nsec;
     } else {
         return 0;
     }
